requirements.txt
alembic==1.17.2
banal==1.0.6
certifi==2025.11.12
charset-normalizer==3.4.4
dataset==1.6.2
discord-webhook==1.4.1
dropbox==12.0.2
greenlet==3.3.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
ply==3.11
pycountry==24.6.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==1.4.54
stone==3.3.1
typing_extensions==4.15.0
urllib3==2.6.2